Create new defaults submodule and move contents of _defaults.py to new yaml file.
