+++ ORBIT/core/components.py
@@ -7,7 +7,7 @@ __email__ = "jake.nunemaker@nrel.gov"
 
 import simpy
 
-from ORBIT.core._defaults import process_times as pt
+from ORBIT.core.defaults import process_times as pt
 from ORBIT.core.exceptions import (
     ItemNotFound,
     CargoMassExceeded,
