+++ ORBIT/core/logic/vessel_logic.py
@@ -8,7 +8,7 @@ __email__ = "jake.nunemaker@nrel.gov"
 
 from marmot import process
 
-from ORBIT.core._defaults import process_times as pt
+from ORBIT.core.defaults import process_times as pt
 from ORBIT.core.exceptions import ItemNotFound, VesselCapacityError
