+++ ORBIT/phases/install/cable_install/common.py
@@ -9,7 +9,7 @@ __email__ = "jake.nunemaker@nrel.gov"
 from marmot import process
 
 from ORBIT.core.logic import position_onsite
-from ORBIT.core._defaults import process_times as pt
+from ORBIT.core.defaults import process_times as pt
 
 
 class SimpleCable:
