+++ ORBIT/phases/install/mooring_install/mooring.py
@@ -10,7 +10,7 @@ from marmot import process
 
 from ORBIT.core import Cargo, Vessel
 from ORBIT.core.logic import position_onsite, get_list_of_items_from_port
-from ORBIT.core._defaults import process_times as pt
+from ORBIT.core.defaults import process_times as pt
 from ORBIT.phases.install import InstallPhase
 from ORBIT.core.exceptions import ItemNotFound
