+++ ORBIT/phases/install/scour_protection_install/standard.py
@@ -12,7 +12,7 @@ import simpy
 from marmot import process
 
 from ORBIT.core import Vessel
-from ORBIT.core._defaults import process_times as pt
+from ORBIT.core.defaults import process_times as pt
 from ORBIT.phases.install import InstallPhase
 from ORBIT.core.exceptions import CargoMassExceeded, InsufficientAmount
