+++ tests/phases/install/monopile_install/test_monopile_install.py
@@ -13,7 +13,7 @@ import pytest
 
 from tests.data import test_weather
 from ORBIT.library import extract_library_specs
-from ORBIT.core._defaults import process_times as pt
+from ORBIT.core.defaults import process_times as pt
 from ORBIT.phases.install import MonopileInstallation
 
 config_wtiv = extract_library_specs("config", "single_wtiv_mono_install")
