+++ tests/phases/install/mooring_install/test_mooring_install.py
@@ -15,7 +15,7 @@ import pytest
 
 from tests.data import test_weather
 from ORBIT.library import extract_library_specs
-from ORBIT.core._defaults import process_times as pt
+from ORBIT.core.defaults import process_times as pt
 from ORBIT.phases.install import MooringSystemInstallation
 
 config = extract_library_specs("config", "mooring_system_install")
