+++ tests/phases/install/scour_protection_install/test_scour_protection.py
@@ -13,7 +13,7 @@ import pytest
 
 from tests.data import test_weather
 from ORBIT.library import extract_library_specs
-from ORBIT.core._defaults import process_times as pt
+from ORBIT.core.defaults import process_times as pt
 from ORBIT.phases.install import ScourProtectionInstallation
 
 config = extract_library_specs("config", "scour_protection_install")
